Implement initial stub of libbfd wrapper C library and test program
We've added a README with some brief thoughts on next steps.

+++ binutils/build-binutils.sh
@@ -1,8 +1,13 @@
 #!/usr/bin/env bash
 
-rm -rf ./binutils-build
-mkdir ./binutils-build
+clean_and_rebuild=false
+
+if [ "$clean_and_rebuild" = true ] ; then
+	rm -rf ./binutils-build
+	mkdir ./binutils-build
+fi
 
 cd ./binutils-build \
 	&& ../binutils-gdb/configure \
 	&& make
+
